+++ tests/Fixtures/Step/no-parameters.yml
@@ -1,0 +1,5 @@
+actions:
+  - click ".button"
+
+assertions:
+  - ".heading" contains "Hello World"
