+++ tests/Unit/Model/Action/UnrecognisedActionTest.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace webignition\BasilParser\Tests\Unit\Model\Action;
+
+use webignition\BasilParser\Model\Action\ActionTypes;
+use webignition\BasilParser\Model\Action\UnrecognisedAction;
+
+class UnrecognisedActionTest extends \PHPUnit\Framework\TestCase
+{
+    public function testCreate()
+    {
+        $type = ActionTypes::RELOAD;
+
+        $action = new UnrecognisedAction($type, '');
+
+        $this->assertSame($type, $action->getType());
+        $this->assertSame('', $action->getArguments());
+        $this->assertFalse($action->isRecognised());
+    }
+}
